testing

+++ src/at/fh/swenga/project/controller/ActivityController.java
@@ -8,6 +8,9 @@ import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
@@ -33,61 +36,73 @@ import at.fh.swenga.project.model.UserRole;
 
 @Controller
 public class ActivityController {
-	String lastcategory; 
-	boolean categoriesCreated = false; 
-	
+	String lastcategory;
+	String currentUser;
+	boolean categoriesCreated = false;
+
 	@Autowired
 	ActivityRepository activityRepository;
-	
+
 	@Autowired
 	SubcategoryRepository subcategoryRepository;
-	
+
 	@Autowired
 	CategoryRepository categoryReposiory;
-	
+
 	@Autowired
 	SimpleUserRepository userRepository;
-	
+
 	@Autowired
 	UserRoleRepository userRoleRepository;
-	
+
 	@RequestMapping(value = { "/" })
-	public String index(Model model){	
-		//if(!categoriesCreated) subcategoryRepository.save(Categories.FillCategories()); //Erstellen aller Catergories + Subcategories; TODO: löschen!!
+	public String index(Model model) {
+		// if(!categoriesCreated)
+		// subcategoryRepository.save(Categories.FillCategories()); //Erstellen
+		// aller Catergories + Subcategories; TODO: löschen!!
 		// auslesen der aktuell eingeloggten person
-		
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        currentUser = auth.getName();
 		categoriesCreated = true;
-	
+
+		model.addAttribute("currentUser", currentUser);
 		return "index";
 	}
-	
-	@RequestMapping(value = "/listActivities") 
-	public String list(Model model, @RequestParam(required=false) String category) {
-		if(category==null) category = lastcategory;
+
+	@RequestMapping(value = "/listActivities")
+	public String list(Model model, @RequestParam(required = false) String category) {
+		if (category == null)
+			category = lastcategory;
 		lastcategory = category;
-	
+
 		List<Activity> activities = (List<Activity>) model.asMap().get("activities");
 		List<Integer> activitiesInt = new ArrayList<>();
-		if(activities!=null) {
-			for(Activity a : activities) { activitiesInt.add(a.getId()); }
+		if (activities != null) {
+			for (Activity a : activities) {
+				activitiesInt.add(a.getId());
+			}
 		}
-		
+
 		List<Subcategory> subcategories = subcategoryRepository.findByCategoryName(category);
-		if(activities==null) activities = activityRepository.getCatActivities(category); // Falls keine Activties übergeben wurden (also keine herausgefiltert wurden)
-		else activities = activityRepository.getFilteredActivities(category, activitiesInt);
-		
+		if (activities == null)
+			activities = activityRepository.getCatActivities(category); // Falls keine Activties übergeben wurden (also keine herausgefiltert wurden)
+		else
+			activities = activityRepository.getFilteredActivities(category, activitiesInt);
+
 		model.addAttribute("activities", activities);
 		model.addAttribute("subcategories", subcategories);
 		model.addAttribute("type", "findAll");
+		model.addAttribute("currentUser", currentUser);
 		return "listActivities";
 	}
 
-	
 	@RequestMapping(value = { "/find" })
-	public String find(Model model, @RequestParam(required=false) String searchString, @RequestParam String type, RedirectAttributes redirectAttributes) { // TODO: Es werden alle passenden angezeigt nicht nur die der jeweiligen Categorie-Liste
+	public String find(Model model, @RequestParam(required = false) String searchString, @RequestParam String type,
+			RedirectAttributes redirectAttributes) { // TODO: Es werden alle passenden angezeigt nicht nur die der jeweiligen Categorie-Liste
 		List<Activity> activities = null;
-		if(searchString==null) type="default"; //Falls kein Filter ausgewählt wurde, werden alle ausgegeben
-		
+		if (searchString == null)
+			type = "default"; // Falls kein Filter ausgewählt wurde, werden alle ausgegeben
+
 		switch (type) {
 		case "findTitle":
 			activities = activityRepository.findByTitleContainingAllIgnoreCase(searchString);
@@ -101,107 +116,115 @@ public class ActivityController {
 		case "findLocation":
 			activities = activityRepository.findByLocationContainingAllIgnoreCase(searchString);
 			break;
-							
+
 		default:
 			activities = activityRepository.findAll();
 		}
 		redirectAttributes.addFlashAttribute("activities", activities);
-		
+
 		return "redirect:listActivities";
 	}
-	
-	
+
 	@RequestMapping("/fill")
 	@Transactional
 	public String fillData(Model model) {
-		
-		
-		Activity a = new Activity(subcategoryRepository.findByName("Soccer"), "Graz", "Steiermark", "Test", "TestText", 1);
+
+		Activity a = new Activity(subcategoryRepository.findByName("Soccer"), "Graz", "Steiermark", "Test", "TestText",
+				1);
 		activityRepository.save(a);
-		
+
 		Activity c = new Activity(subcategoryRepository.findByName("Tennis"), "Wien", "Wien", "ka", "KaText", 1);
 		activityRepository.save(c);
-		
-		Activity b = new Activity(subcategoryRepository.findByName("Counter Strike"), "Graz","Steiermark", "Test", "TestText", 1);
+
+		Activity b = new Activity(subcategoryRepository.findByName("Counter Strike"), "Graz", "Steiermark", "Test",
+				"TestText", 1);
 		activityRepository.save(b);
-	
-		
+
 		return "forward:listActivities";
 	}
-	
+
 	@RequestMapping("/addActivity")
 	public String addActivity(Model model) {
-		
+
 		List<Subcategory> sub = subcategoryRepository.findAll(); // TODO: Nur die richtigen Anzeigen
 		model.addAttribute("subcategories", sub);
 		return "addActivities";
 	}
-	
+
 	@RequestMapping("/fullActivity")
-	public String fullActivity(Model model, @RequestParam(required=false) int id) {
+	public String fullActivity(Model model, @RequestParam(required = false) int id) {
 		Activity a = activityRepository.findById(id);
 		model.addAttribute("activity", a);
-		//model.addAttribute("subcategories", sub);
+		model.addAttribute("currentUser", currentUser);
 		return "activity";
 	}
-	
+
 	@RequestMapping("/add")
-	public String addActivityInDatabase(Model model, @RequestParam String title, @RequestParam String text, @RequestParam String state, @RequestParam String location, @RequestParam @DateTimeFormat(pattern = "dd.MM.yyyy") Date date, @RequestParam int restriction, @RequestParam String type, @RequestParam(required=false) boolean closed  ) {
-		Subcategory s = subcategoryRepository.findByName(type); // TODO: Sonst Error
-		Activity a = new Activity(s, location ,state, title, date,  text, restriction, closed);
+	public String addActivityInDatabase(Model model, @RequestParam String title, @RequestParam String text,
+			@RequestParam String state, @RequestParam String location,
+			@RequestParam @DateTimeFormat(pattern = "dd.MM.yyyy") Date date, @RequestParam int restriction,
+			@RequestParam String type, @RequestParam(required = false) boolean closed) {
+		Subcategory s = subcategoryRepository.findByName(type); // TODO: Sonst
+																// Error
+		Activity a = new Activity(s, location, state, title, date, text, restriction, closed);
 		activityRepository.save(a);
-		
+
 		return "forward:listActivities";
 	}
-	
+
 	@RequestMapping("/user")
 	public String showUserProfile(Model model) {
 		return "showUserProfile";
 	}
 
+//	@PreAuthorize("#contact.name == authentication.name") Vergleiche übergebene ID mit 
 	@RequestMapping("/delete")
 	public String deleteData(Model model, @RequestParam int id) {
-		activityRepository.delete(id);
-		model.addAttribute("warningMessage", "Activity " + id + " deleted");
-
+		Activity a = activityRepository.findById(id);
+		if(a.getOwner().getUsername().equals(currentUser)){
+			activityRepository.delete(id);
+			model.addAttribute("warningMessage", "Activity " + id + " deleted");
+		}
+		else
+			model.addAttribute("warningMessage", "Activity " + id + "not deleted");
+		
 		return "forward:listActivities";
 	}
 
-	//@ExceptionHandler(Exception.class) TODO: Wieder aktivieren nach fertigstellung
+	// @ExceptionHandler(Exception.class) TODO: Wieder aktivieren nach
+	// fertigstellung
 	public String handleAllException(Exception ex) {
 		return "showError";
 	}
-	
+
 	@RequestMapping("/registrationForm")
-	public String registration()
-	{
+	public String registration() {
 		return "registrationForm";
 	}
-	
+
 	@RequestMapping(value = "/login", method = RequestMethod.GET)
 	public String handleLogin() {
-		
+
 		return "login";
 	}
-	
+
 	@RequestMapping("/registrate")
-	public String registrateUser(Model model, @RequestParam String username, @RequestParam String password, @RequestParam String name, @RequestParam String age, @RequestParam String city)
-	{
-		
-		String pw_hash = BCrypt.hashpw(password, BCrypt.gensalt()); 
+	public String registrateUser(Model model, @RequestParam String username, @RequestParam String password,
+			@RequestParam String name, @RequestParam String age, @RequestParam String city) {
+
+		String pw_hash = BCrypt.hashpw(password, BCrypt.gensalt());
 		Set<UserRole> userRole = new HashSet<UserRole>();
-		
+
 		User u = new User(username, pw_hash, true, name, Integer.parseInt(age), city);
 		userRepository.save(u);
-		
+
 		UserRole role = new UserRole(u, "ROLE_USER");
 		userRoleRepository.save(role);
-		
+
 		userRole.add(role);
 		u.addRole(userRole);
-		
+
 		return "login";
 	}
-	
-	
+
 }
